+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.2.0] - 2023-09-13
+
 ### Added
 
 - Add an utils function sign_parameters used for testing
@@ -49,7 +51,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Import `lti_toolbox` library and tests from
   [Ashley](https://github.com/openfun/ashley)
 
-[Unreleased]: https://github.com/openfun/django-lti-toolbox/compare/v1.1.0...master
+[Unreleased]: https://github.com/openfun/django-lti-toolbox/compare/v1.2.0...master
+[1.2.0]: https://github.com/openfun/django-lti-toolbox/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/openfun/django-lti-toolbox/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/openfun/django-lti-toolbox/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/openfun/django-lti-toolbox/compare/v1.0.0b1...v1.0.0